Fix comparing signed/unsigned within 7zHandlerOut.cpp

+++ CPP/7zip/Archive/7z/7zHandlerOut.cpp
@@ -453,7 +453,8 @@ Z7_COM7F_IMF(CHandler::UpdateItems(ISequentialOutStream *outStream, UInt32 numIt
       if (!mInfo.methName.IsEmpty()) {
         COneMethodInfo &m = _methods.AddNew();
         m.MethodName = mInfo.methName;
-        if ((_level == -1) && (mInfo.level != -1)) {
+        if ((_level == (UInt32)-1) && (mInfo.level != -1)) {
+          // XXX - does not work in linux this way
           _level = mInfo.level;
         }
       }
